introducing character_sequence
For the matcher we need to know the current character in the input, but
also where we are in the input - at start, at end, in the middle.
character_sequence will wrap that up together so we don't have to pass
lots of separate things around.

This commit sketches out the start of the interface.