special handling for terminal-only patterns (%|$)
I'm not super keen to do this, but it turned out we were not matching a
naked $ with the existing set up. It's always going to be a bit weird,
and just pulling out terminal-only patterns seems neater than trying to
futz with the core loop. As a minor side-effect, not that youd notice,
quicker too.