edit: use from_is_good to tidy up action creation
I'm not entirely sure I'm wedded to this, but cleaning up the early
return feels nice. I guess we could work this through with std::optional

  optional(from_index != 0).transform(action).value_or(error_action)

but that feels like the long way round to the same destination.
Oh wait, hang on, optional doesn't work like that. I'm thinking of
something more akin to a Javascript promise or something.