format: poked around with tests again
Drop test input and expectations as files in the test_fixture
subdirectory. Walk the directory looking for .txt files, use them as
input the formatter, compare against <same name>.exp.

This is the kind of thing I don't usually do in C++ even though I'm
entirely fine with it in JavaScript. Go figure.

Generated sample outputs using nroff, because it still has some of the
same assumptions embedded in it. Weird, huh.

Anyway, the tests run (at least using CLion's embedded text runner), but
they are entirely broken in this commit. I feel a bit naughty doing
that, but it's bedtime rn