format: rework for mid-line underlining
The underline control codes make the string longer, but not the
displayed width. Consequently, when working out the centering offset or
if the line has wrapped, we need to ignore any underlining. This change
replaces calls to line.size() with calls to count_width(line). When
filling, we now buffer a word at a time and flush if required, rather
than buffering each line, then splitting the whole buffer.

Sounds like a lot, but it's a less radical change than you might think.